+++ scripts/deno/mod.ts
@@ -120,11 +120,11 @@ export const DEFAULT_CONFIGURATION: Configuration = {
 
 const DOWNLOAD_URLS: Record<typeof Deno.build.os, string> = {
   "linux":
-    "https://github.com/dahlia/seonbi/releases/download/0.2.1/seonbi-0.2.1.linux-x86_64.tar.bz2",
+    "https://github.com/dahlia/seonbi/releases/download/0.2.2/seonbi-0.2.2.linux-x86_64.tar.bz2",
   "darwin":
-    "https://github.com/dahlia/seonbi/releases/download/0.2.1/seonbi-0.2.1.macos-x86_64.tar.bz2",
+    "https://github.com/dahlia/seonbi/releases/download/0.2.2/seonbi-0.2.2.macos-x86_64.tar.bz2",
   "windows":
-    "https://github.com/dahlia/seonbi/releases/download/0.2.1/seonbi-0.2.1.win64.zip",
+    "https://github.com/dahlia/seonbi/releases/download/0.2.2/seonbi-0.2.2.win64.zip",
 } as const;
 
 /**